Divide class that holds settings for customizing default behavior
