+++ src/main/java/org/apache/ibatis/session/Settings.java
@@ -1,0 +1,333 @@
+/**
+ *    Copyright 2009-2017 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.apache.ibatis.session;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+import org.apache.ibatis.executor.loader.ProxyFactory;
+import org.apache.ibatis.executor.loader.javassist.JavassistProxyFactory;
+import org.apache.ibatis.io.VFS;
+import org.apache.ibatis.logging.Log;
+import org.apache.ibatis.logging.LogFactory;
+import org.apache.ibatis.scripting.xmltags.XMLLanguageDriver;
+import org.apache.ibatis.type.EnumTypeHandler;
+import org.apache.ibatis.type.JdbcType;
+import org.apache.ibatis.type.TypeHandler;
+
+/**
+ * Holds settings for customizing default behavior of MyBatis.
+ * <p>
+ *   This class was created by separating the {@link Configuration}.
+ * </p>
+ *
+ * @author Clinton Begin
+ * @author Kazuki Shimizu
+ * @since 3.4.6
+ */
+public class Settings {
+
+  protected boolean safeRowBoundsEnabled;
+  protected boolean safeResultHandlerEnabled = true;
+  protected boolean mapUnderscoreToCamelCase;
+  protected boolean aggressiveLazyLoading;
+  protected boolean multipleResultSetsEnabled = true;
+  protected boolean useGeneratedKeys;
+  protected boolean useColumnLabel = true;
+  protected boolean cacheEnabled = true;
+  protected boolean callSettersOnNulls;
+  protected boolean useActualParamName = true;
+  protected boolean returnInstanceForEmptyRow;
+
+  protected String logPrefix;
+  protected Class <? extends Log> logImpl;
+  protected Class <? extends VFS> vfsImpl;
+  protected LocalCacheScope localCacheScope = LocalCacheScope.SESSION;
+  protected JdbcType jdbcTypeForNull = JdbcType.OTHER;
+  protected Set<String> lazyLoadTriggerMethods = new HashSet<String>(Arrays.asList(new String[] { "equals", "clone", "hashCode", "toString" }));
+  protected Integer defaultStatementTimeout;
+  protected Integer defaultFetchSize;
+  protected ExecutorType defaultExecutorType = ExecutorType.SIMPLE;
+  protected AutoMappingBehavior autoMappingBehavior = AutoMappingBehavior.PARTIAL;
+  protected AutoMappingUnknownColumnBehavior autoMappingUnknownColumnBehavior = AutoMappingUnknownColumnBehavior.NONE;
+
+  protected boolean lazyLoadingEnabled = false;
+  protected ProxyFactory proxyFactory = new JavassistProxyFactory(); // #224 Using internal Javassist instead of OGNL
+
+  protected Class<? extends TypeHandler> defaultEnumTypeHandler = EnumTypeHandler.class;
+  protected Class<?> defaultScriptingLanguage;
+
+  /**
+   * Configuration factory class.
+   * Used to create Configuration for loading deserialized unread properties.
+   *
+   * @see <a href='https://code.google.com/p/mybatis/issues/detail?id=300'>Issue 300 (google code)</a>
+   */
+  protected Class<?> configurationFactory;
+
+  public String getLogPrefix() {
+    return logPrefix;
+  }
+
+  public void setLogPrefix(String logPrefix) {
+    this.logPrefix = logPrefix;
+  }
+
+  public Class<? extends Log> getLogImpl() {
+    return logImpl;
+  }
+
+  public void setLogImpl(Class<? extends Log> logImpl) {
+    if (logImpl != null) {
+      this.logImpl = logImpl;
+      LogFactory.useCustomLogging(this.logImpl);
+    }
+  }
+
+  public Class<? extends VFS> getVfsImpl() {
+    return this.vfsImpl;
+  }
+
+  public void setVfsImpl(Class<? extends VFS> vfsImpl) {
+    if (vfsImpl != null) {
+      this.vfsImpl = vfsImpl;
+      VFS.addImplClass(this.vfsImpl);
+    }
+  }
+
+  public boolean isCallSettersOnNulls() {
+    return callSettersOnNulls;
+  }
+
+  public void setCallSettersOnNulls(boolean callSettersOnNulls) {
+    this.callSettersOnNulls = callSettersOnNulls;
+  }
+
+  public boolean isUseActualParamName() {
+    return useActualParamName;
+  }
+
+  public void setUseActualParamName(boolean useActualParamName) {
+    this.useActualParamName = useActualParamName;
+  }
+
+  public boolean isReturnInstanceForEmptyRow() {
+    return returnInstanceForEmptyRow;
+  }
+
+  public void setReturnInstanceForEmptyRow(boolean returnEmptyInstance) {
+    this.returnInstanceForEmptyRow = returnEmptyInstance;
+  }
+
+  public Class<?> getConfigurationFactory() {
+    return configurationFactory;
+  }
+
+  public void setConfigurationFactory(Class<?> configurationFactory) {
+    this.configurationFactory = configurationFactory;
+  }
+
+  public boolean isSafeResultHandlerEnabled() {
+    return safeResultHandlerEnabled;
+  }
+
+  public void setSafeResultHandlerEnabled(boolean safeResultHandlerEnabled) {
+    this.safeResultHandlerEnabled = safeResultHandlerEnabled;
+  }
+
+  public boolean isSafeRowBoundsEnabled() {
+    return safeRowBoundsEnabled;
+  }
+
+  public void setSafeRowBoundsEnabled(boolean safeRowBoundsEnabled) {
+    this.safeRowBoundsEnabled = safeRowBoundsEnabled;
+  }
+
+  public boolean isMapUnderscoreToCamelCase() {
+    return mapUnderscoreToCamelCase;
+  }
+
+  public void setMapUnderscoreToCamelCase(boolean mapUnderscoreToCamelCase) {
+    this.mapUnderscoreToCamelCase = mapUnderscoreToCamelCase;
+  }
+
+  public AutoMappingBehavior getAutoMappingBehavior() {
+    return autoMappingBehavior;
+  }
+
+  public void setAutoMappingBehavior(AutoMappingBehavior autoMappingBehavior) {
+    this.autoMappingBehavior = autoMappingBehavior;
+  }
+
+  /**
+   * @since 3.4.0
+   */
+  public AutoMappingUnknownColumnBehavior getAutoMappingUnknownColumnBehavior() {
+    return autoMappingUnknownColumnBehavior;
+  }
+
+  /**
+   * @since 3.4.0
+   */
+  public void setAutoMappingUnknownColumnBehavior(AutoMappingUnknownColumnBehavior autoMappingUnknownColumnBehavior) {
+    this.autoMappingUnknownColumnBehavior = autoMappingUnknownColumnBehavior;
+  }
+
+  public boolean isLazyLoadingEnabled() {
+    return lazyLoadingEnabled;
+  }
+
+  public void setLazyLoadingEnabled(boolean lazyLoadingEnabled) {
+    this.lazyLoadingEnabled = lazyLoadingEnabled;
+  }
+
+  public ProxyFactory getProxyFactory() {
+    return proxyFactory;
+  }
+
+  public void setProxyFactory(ProxyFactory proxyFactory) {
+    if (proxyFactory == null) {
+      proxyFactory = new JavassistProxyFactory();
+    }
+    this.proxyFactory = proxyFactory;
+  }
+
+  public boolean isAggressiveLazyLoading() {
+    return aggressiveLazyLoading;
+  }
+
+  public void setAggressiveLazyLoading(boolean aggressiveLazyLoading) {
+    this.aggressiveLazyLoading = aggressiveLazyLoading;
+  }
+
+  public boolean isMultipleResultSetsEnabled() {
+    return multipleResultSetsEnabled;
+  }
+
+  public void setMultipleResultSetsEnabled(boolean multipleResultSetsEnabled) {
+    this.multipleResultSetsEnabled = multipleResultSetsEnabled;
+  }
+
+  public Set<String> getLazyLoadTriggerMethods() {
+    return lazyLoadTriggerMethods;
+  }
+
+  public void setLazyLoadTriggerMethods(Set<String> lazyLoadTriggerMethods) {
+    this.lazyLoadTriggerMethods = lazyLoadTriggerMethods;
+  }
+
+  public boolean isUseGeneratedKeys() {
+    return useGeneratedKeys;
+  }
+
+  public void setUseGeneratedKeys(boolean useGeneratedKeys) {
+    this.useGeneratedKeys = useGeneratedKeys;
+  }
+
+  public ExecutorType getDefaultExecutorType() {
+    return defaultExecutorType;
+  }
+
+  public void setDefaultExecutorType(ExecutorType defaultExecutorType) {
+    this.defaultExecutorType = defaultExecutorType;
+  }
+
+  public boolean isCacheEnabled() {
+    return cacheEnabled;
+  }
+
+  public void setCacheEnabled(boolean cacheEnabled) {
+    this.cacheEnabled = cacheEnabled;
+  }
+
+  public Integer getDefaultStatementTimeout() {
+    return defaultStatementTimeout;
+  }
+
+  public void setDefaultStatementTimeout(Integer defaultStatementTimeout) {
+    this.defaultStatementTimeout = defaultStatementTimeout;
+  }
+
+  /**
+   * @since 3.3.0
+   */
+  public Integer getDefaultFetchSize() {
+    return defaultFetchSize;
+  }
+
+  /**
+   * @since 3.3.0
+   */
+  public void setDefaultFetchSize(Integer defaultFetchSize) {
+    this.defaultFetchSize = defaultFetchSize;
+  }
+
+  public boolean isUseColumnLabel() {
+    return useColumnLabel;
+  }
+
+  public void setUseColumnLabel(boolean useColumnLabel) {
+    this.useColumnLabel = useColumnLabel;
+  }
+
+  public LocalCacheScope getLocalCacheScope() {
+    return localCacheScope;
+  }
+
+  public void setLocalCacheScope(LocalCacheScope localCacheScope) {
+    this.localCacheScope = localCacheScope;
+  }
+
+  public JdbcType getJdbcTypeForNull() {
+    return jdbcTypeForNull;
+  }
+
+  public void setJdbcTypeForNull(JdbcType jdbcTypeForNull) {
+    this.jdbcTypeForNull = jdbcTypeForNull;
+  }
+
+  /**
+   * Set a default {@link TypeHandler} class for {@link Enum}.
+   * A default {@link TypeHandler} is {@link org.apache.ibatis.type.EnumTypeHandler}.
+   * @param typeHandler a type handler class for {@link Enum}
+   * @since 3.4.5
+   */
+  public void setDefaultEnumTypeHandler(Class<? extends TypeHandler> typeHandler) {
+    if (typeHandler != null) {
+      this.defaultEnumTypeHandler = typeHandler;
+    }
+  }
+
+  /**
+   * @since 3.4.6
+   */
+  public Class<? extends TypeHandler> getDefaultEnumTypeHandler() {
+    return defaultEnumTypeHandler;
+  }
+
+  public void setDefaultScriptingLanguage(Class<?> driver) {
+    this.defaultScriptingLanguage = driver != null ? driver : XMLLanguageDriver.class;
+  }
+
+  /**
+   * @since 3.4.6
+   */
+  public Class<?> getDefaultScriptingLanguage() {
+    return defaultScriptingLanguage;
+  }
+
+}
